feat: configure routing and navigation entry point
added app entry point, moved home from root to main
