+++ app/(main)/home/_layout.jsx
@@ -1,0 +1,14 @@
+import { Stack } from "expo-router";
+
+export default function HomeLayout() {
+  return (
+    <Stack>
+      <Stack.Screen
+        name="index"
+        options={{
+          headerShown: false,
+        }}
+      />
+    </Stack>
+  );
+}
